cleanup cloudwatch.go
